Proof of concept
Will start at the provided root-server and make an NS-query to find next
step. Using the addresses in the additional-section it will then go to
the next label until reaching the goal.

For each step we query the parent-zone for DS-records and the child-zone
for DNSKEY-records and compare key-tag and algorithm to see if it
matches.